Pipeline input refactor - finished
Concept of vertex arrays eliminated from the public API, components now bound to the active graphics pipeline individually in line with modern graphics API's.

Vertex array still created and cached in background. The additional lookup adds some additional overhead as compared to the prior situation, but this change sets up safe borrowed checked transform feedback sessions (buffer used to record feedback may not also be bound to the pipeline as input), which was not possible in the vertex array situation (at least not without adding lifetimes for each buffer to the vertex array, which would not really have been compatible with something intended to be a long lived object).